who/requirements.txt
django
south
django-piston
